refactor getRepository, so we can use Upgrade/Install outside of setup too
