+++ packages/cli/src/lib/setup/setupUpgrade.ts
@@ -17,7 +17,8 @@ import tty from 'tty';
 import path from 'path';
 import type { Client as ObjectsInRedisClient } from '@iobroker/db-objects-redis';
 import type { Client as StatesInRedisClient } from '@iobroker/db-states-redis';
-import type { GetRepositoryHandler, ProcessExitCallback } from '../_Types';
+import type { ProcessExitCallback } from '../_Types';
+import { getRepository } from './utils';
 
 const debug = Debug('iobroker:cli');
 
@@ -26,7 +27,6 @@ type IoPackDependencies = string[] | Record<string, any>[] | Record<string, any>
 interface CLIUpgradeOptions {
     processExit: ProcessExitCallback;
     restartController: () => void;
-    getRepository: GetRepositoryHandler;
     objects: ObjectsInRedisClient;
     states: StatesInRedisClient;
     params: Record<string, any>;
@@ -39,7 +39,6 @@ export class Upgrade {
     private objects: ObjectsInRedisClient;
     private readonly processExit: ProcessExitCallback;
     private readonly params: Record<string, any>;
-    private readonly getRepository: GetRepositoryHandler;
 
     constructor(options: CLIUpgradeOptions) {
         options = options || {};
@@ -50,12 +49,8 @@ export class Upgrade {
         if (!options.restartController) {
             throw new Error('Invalid arguments: restartController is missing');
         }
-        if (!options.getRepository) {
-            throw new Error('Invalid arguments: getRepository is missing');
-        }
 
         this.processExit = options.processExit;
-        this.getRepository = options.getRepository;
         this.params = options.params;
         this.objects = options.objects;
 
@@ -314,7 +309,7 @@ export class Upgrade {
         let sources: Record<string, any>;
         if (!repoUrlOrObject || !tools.isObject(repoUrlOrObject)) {
             try {
-                sources = await this.getRepository(repoUrlOrObject, this.params);
+                sources = await getRepository(this.objects, repoUrlOrObject);
             } catch (e) {
                 return this.processExit(e);
             }
@@ -649,7 +644,7 @@ export class Upgrade {
         let sources: Record<string, any>;
         if (!repoUrlOrObject || !tools.isObject(repoUrlOrObject)) {
             try {
-                const result = await this.getRepository(repoUrlOrObject, this.params);
+                const result = await getRepository(this.objects, repoUrlOrObject);
                 if (!result) {
                     return console.warn(`Cannot get repository under "${repoUrlOrObject}"`);
                 }
